+++ samples/CustomV3InProcessFunction/Functions/TriggerServiceBusExceptionThrowing/TriggerServiceBusExceptionThrowingFunction.cs
