+++ samples/DefaultV3InProcessFunction/Functions/UserSecret/SecretOptions.cs
@@ -1,0 +1,6 @@
+namespace Gabo.AzureFunctionTelemetry.Samples.DefaultV3InProcessFunction.Functions.UserSecret;
+
+public class SecretOptions
+{
+    public string ReallySecretValue { get; set; } = default!;
+}
