+++ samples/DefaultV4InProcessFunction/Functions/TriggerServiceBusExceptionThrowing/TriggerServiceBusExceptionThrowingFunction.cs
