+++ tests/AppInsightsConnectionStringIntegrationTests/TestInfrastructure/CustomTelemetryFunctionClient.cs
@@ -1,0 +1,8 @@
+namespace Gabo.AzureFunctionsTelemetryAppInsightsConnectionStringIntegrationTests.TestInfrastructure;
+
+internal sealed class CustomTelemetryFunctionClient : TelemetryFunctionClient
+{
+    public CustomTelemetryFunctionClient() : base(7074)
+    {
+    }
+}
