+++ tests/AppInsightsConnectionStringIntegrationTests/TestInfrastructure/DefaultTelemetryFunctionClient.cs
@@ -1,0 +1,8 @@
+namespace Gabo.AzureFunctionsTelemetryAppInsightsConnectionStringIntegrationTests.TestInfrastructure;
+
+internal sealed class DefaultTelemetryFunctionClient : TelemetryFunctionClient
+{
+    public DefaultTelemetryFunctionClient() : base(7073)
+    {
+    }
+}
