+++ tests/AzureFunctionsTelemetryIntegrationTests/AzureFunctionsTelemetryIntegrationTests.csproj
@@ -8,7 +8,7 @@
     <RootNamespace>Gabo.AzureFunctionsTelemetryIntegrationTests</RootNamespace>
     <AssemblyName>Gabo.AzureFunctionsTelemetryIntegrationTests</AssemblyName>
     <IsPackable>false</IsPackable>
-    <NoWarn>7035;CA1707;CA2007</NoWarn>
+    <NoWarn>7035;CA1707;CA2007;MA0004</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
