+++ tests/AzureFunctionsTelemetryIntegrationTests/TestInfrastructure/DefaultTelemetryFunctionClient.cs
@@ -1,0 +1,8 @@
+namespace Gabo.AzureFunctionsTelemetryIntegrationTests.TestInfrastructure;
+
+internal sealed class DefaultTelemetryFunctionClient : TelemetryFunctionClient
+{
+    public DefaultTelemetryFunctionClient() : base(7073)
+    {
+    }
+}
