+++ tests/AzureFunctionsTelemetryIntegrationTests/TestInfrastructure/RequestData.cs
@@ -1,0 +1,6 @@
+namespace Gabo.AzureFunctionsTelemetryIntegrationTests.TestInfrastructure;
+
+internal sealed class RequestData
+{
+    public RequestBaseData BaseData { get; set; } = new();
+}
