+++ tests/AzureFunctionsTelemetryIntegrationTests/TestInfrastructure/TraceBaseData.cs
@@ -1,0 +1,6 @@
+namespace Gabo.AzureFunctionsTelemetryIntegrationTests.TestInfrastructure;
+
+internal sealed class TraceBaseData
+{
+    public string Message { get; set; } = default!;
+}
