+++ tests/AzureFunctionsTelemetryIntegrationTests/TestInfrastructure/TraceData.cs
@@ -1,0 +1,6 @@
+namespace Gabo.AzureFunctionsTelemetryIntegrationTests.TestInfrastructure;
+
+internal sealed class TraceData
+{
+    public TraceBaseData BaseData { get; set; } = new();
+}
