+++ tests/AzureFunctionsTelemetryIntegrationTests/TestInfrastructure/TraceItem.cs
@@ -4,13 +4,3 @@ internal sealed class TraceItem : TelemetryItem
 {
     public TraceData Data { get; set; } = new();
 }
-
-internal sealed class TraceData
-{
-    public TraceBaseData BaseData { get; set; } = new();
-}
-
-internal sealed class TraceBaseData
-{
-    public string Message { get; set; } = default!;
-}
