+++ tests/AzureFunctionsTelemetryTests/ApplicationInsights/FunctionsFinderTests.cs
@@ -1,8 +1,4 @@
-using System.Diagnostics.CodeAnalysis;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
+using Gabo.AzureFunctionsTelemetryTests.TestInfrastructure.Functions;
 
 namespace Gabo.AzureFunctionsTelemetryTests.ApplicationInsights;
 
@@ -35,37 +31,4 @@ public class FunctionsFinderTests
         // Assert
         _foundFunctions.Should().NotContain(nameof(HttpFunction));
     }
-}
-
-public static class StaticServiceBusFunction
-{
-    [FunctionName(nameof(StaticServiceBusFunction))]
-    public static void Run(
-        [ServiceBusTrigger("static-queue", Connection = "ServiceBusConnection")]
-        [SuppressMessage("Style", "IDE0060", Justification = "Can't use discard as it breaks the binding")]
-        string myQueueItem)
-    {
-    }
-}
-
-public static class InstanceServiceBusFunction
-{
-    [FunctionName(nameof(InstanceServiceBusFunction))]
-    public static void Run(
-        [ServiceBusTrigger("instance-queue", Connection = "ServiceBusConnection")]
-        [SuppressMessage("Style", "IDE0060", Justification = "Can't use discard as it breaks the binding")]
-        string myQueueItem)
-    {
-    }
-}
-
-public static class HttpFunction
-{
-    [HttpGet]
-    [FunctionName(nameof(HttpFunction))]
-    public static IActionResult Run(
-        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "http")]
-        [SuppressMessage("Style", "IDE0060", Justification = "Can't use discard as it breaks the binding")]
-        HttpRequest request) =>
-        new OkResult();
 }
