+++ tests/AzureFunctionsTelemetryTests/TestInfrastructure/Functions/HttpFunction.cs
@@ -1,0 +1,18 @@
+using System.Diagnostics.CodeAnalysis;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Extensions.Http;
+
+namespace Gabo.AzureFunctionsTelemetryTests.TestInfrastructure.Functions;
+
+public static class HttpFunction
+{
+    [HttpGet]
+    [FunctionName(nameof(HttpFunction))]
+    public static IActionResult Run(
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "http")]
+        [SuppressMessage("Style", "IDE0060", Justification = "Can't use discard as it breaks the binding")]
+        HttpRequest request) =>
+        new OkResult();
+}
